requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kaitaistruct==0.10
lxml==4.9.1
outcome==1.2.0
pyasn1==0.4.8
pycparser==2.21
pymongo==4.2.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
selenium==4.4.3
selenium-wire==4.6.5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
zstandard==0.18.0
